# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mini Admin Dashboard API",
    "description": "OpenAPI specification for Mini Admin Dashboard",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/api/products": {
      "get": {
        "summary": "Get products list",
        "description": "Retrieve paginated list of products",
        "responses": {
          "200": {
            "description": "List of products"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create product",
        "description": "Create a new product (admin only)",
        "responses": {
          "201": {
            "description": "Product created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/products/{id}": {
      "patch": {
        "summary": "Update product",
        "description": "Update product by ID (admin only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete product",
        "description": "Soft delete product (admin only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  }
}